# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "http server api docs",
        "title": "helmsman api docs",
        "contact": {},
        "version": "v1.0.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/trades": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new trades entity using the provided data in the request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trades"
                ],
                "summary": "Create a new trades",
                "parameters": [
                    {
                        "description": "trades information",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateTradesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTradesReply"
                        }
                    }
                }
            }
        },
        "/api/v1/trades/list": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of trades based on query filters, including page number and size.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trades"
                ],
                "summary": "Get a paginated list of tradess by custom conditions",
                "parameters": [
                    {
                        "description": "query parameters",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Params"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ListTradessReply"
                        }
                    }
                }
            }
        },
        "/api/v1/trades/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets detailed information of a trades specified by the given id in the path.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trades"
                ],
                "summary": "Get a trades by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetTradesByIDReply"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the specified trades by given id in the path, support partial update.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trades"
                ],
                "summary": "Update a trades by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "trades information",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateTradesByIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UpdateTradesByIDReply"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a existing trades identified by the given id in the path.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trades"
                ],
                "summary": "Delete a trades by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.DeleteTradesByIDReply"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.Column": {
            "type": "object",
            "properties": {
                "exp": {
                    "description": "expressions, which default to = when the value is null, have =, !=, \u003e, \u003e=, \u003c, \u003c=, like",
                    "type": "string"
                },
                "logic": {
                    "description": "logical type, default value is \"and\", support \u0026, and, ||, or",
                    "type": "string"
                },
                "name": {
                    "description": "column name",
                    "type": "string"
                },
                "value": {
                    "description": "column value"
                }
            }
        },
...
        "types.CreateTradesReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "id": {
                            "description": "id",
                            "type": "integer"
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
        "types.CreateTradesRequest": {
            "type": "object",
            "properties": {
                "accountID": {
                    "type": "integer"
                },
                "actualEntryPrice": {
                    "type": "number"
                },
                "actualEntryTime": {
                    "type": "string"
                },
                "actualExitPrice": {
                    "type": "number"
                },
                "actualExitTime": {
                    "type": "string"
                },
                "commission": {
                    "type": "number"
                },
                "direction": {
                    "type": "string"
                },
                "executionScore": {
                    "type": "integer"
                },
                "exitReason": {
                    "type": "string"
                },
                "planNotes": {
                    "type": "string"
                },
                "plannedEntryPrice": {
                    "type": "number"
                },
                "plannedRiskAmount": {
                    "type": "number"
                },
                "plannedStopLoss": {
                    "type": "number"
                },
                "plannedTakeProfit": {
                    "type": "number"
                },
                "pnl": {
                    "type": "number"
                },
                "positionSize": {
                    "type": "number"
                },
                "rMultiple": {
                    "type": "number"
                },
                "reflectionNotes": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "strategyID": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
...
        "types.DeleteTradesByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object"
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.GetTradesByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "trades": {
                            "$ref": "#/definitions/types.TradesObjDetail"
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.ListTradessReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "tradess": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.TradesObjDetail"
                            }
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.Params": {
            "type": "object",
            "properties": {
                "columns": {
                    "description": "query conditions",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Column"
                    }
                },
                "limit": {
                    "description": "lines per page",
                    "type": "integer"
                },
                "page": {
                    "description": "page number, starting from page 0",
                    "type": "integer"
                },
                "sort": {
                    "description": "sorted fields, multi-column sorting separated by commas",
                    "type": "string"
                }
            }
        },
...
        "types.TradesObjDetail": {
            "type": "object",
            "properties": {
                "accountID": {
                    "type": "integer"
                },
                "actualEntryPrice": {
                    "type": "number"
                },
                "actualEntryTime": {
                    "type": "string"
                },
                "actualExitPrice": {
                    "type": "number"
                },
                "actualExitTime": {
                    "type": "string"
                },
                "commission": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "direction": {
                    "type": "string"
                },
                "executionScore": {
                    "type": "integer"
                },
                "exitReason": {
                    "type": "string"
                },
                "id": {
                    "description": "convert to uint64 id",
                    "type": "integer"
                },
                "planNotes": {
                    "type": "string"
                },
                "plannedEntryPrice": {
                    "type": "number"
                },
                "plannedRiskAmount": {
                    "type": "number"
                },
                "plannedStopLoss": {
                    "type": "number"
                },
                "plannedTakeProfit": {
                    "type": "number"
                },
                "pnl": {
                    "type": "number"
                },
                "positionSize": {
                    "type": "number"
                },
                "rMultiple": {
                    "type": "number"
                },
                "reflectionNotes": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "strategyID": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "types.UpdateTradesByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object"
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
        "types.UpdateTradesByIDRequest": {
            "type": "object",
            "properties": {
                "accountID": {
                    "type": "integer"
                },
                "actualEntryPrice": {
                    "type": "number"
                },
                "actualEntryTime": {
                    "type": "string"
                },
                "actualExitPrice": {
                    "type": "number"
                },
                "actualExitTime": {
                    "type": "string"
                },
                "commission": {
                    "type": "number"
                },
                "direction": {
                    "type": "string"
                },
                "executionScore": {
                    "type": "integer"
                },
                "exitReason": {
                    "type": "string"
                },
                "id": {
                    "description": "uint64 id",
                    "type": "integer"
                },
                "planNotes": {
                    "type": "string"
                },
                "plannedEntryPrice": {
                    "type": "number"
                },
                "plannedRiskAmount": {
                    "type": "number"
                },
                "plannedStopLoss": {
                    "type": "number"
                },
                "plannedTakeProfit": {
                    "type": "number"
                },
                "pnl": {
                    "type": "number"
                },
                "positionSize": {
                    "type": "number"
                },
                "rMultiple": {
                    "type": "number"
                },
                "reflectionNotes": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "strategyID": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type Bearer your-jwt-token to Value",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}